リファクタリング　モジュール分散
１ファイルにまとめていたクラスや初期読み込み後処理をモジュールに分散
