+++ main/index.html
@@ -15,6 +15,6 @@
     <h2>Picture Information Viewer</h2>
     <div id="pictures-dir" class="button button-outline">写真格納フォルダの選択</div>
     <div id="pictures-list" class="container"></div>
-    <script src="../renderer/model/choose-pictures-folder.js"></script>
+    <script src="../renderer/model/dom-content-loaded.js"></script>
   </body>
 </html>
