+++ renderer/model/dom-content-loaded.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const PicturesDir = require('../renderer/model/pictures-dir');
+
+document.addEventListener('DOMContentLoaded', () => {
+  new PicturesDir();
+});
