+++ renderer/model/pictures-list.js
@@ -1,61 +1,6 @@
 'use strict';
 
-const { remote, ipcRenderer, clipboard } = require('electron');
-const Dialog = remote.dialog;
-const g = {};
-
-
-document.addEventListener('DOMContentLoaded', () => {
-  g.picturesDir = new PicturesDir();
-  g.picturesList = new PicturesList();
-});
-
-
-
-class PicturesDir {
-
-  constructor() {
-    this.selector = document.querySelector('#pictures-dir');
-    this.selector.addEventListener('click', (arg) => this.onClick(arg));
-    this.ipcReply();
-  }
-
-  /** @private */
-  onClick() {
-    Dialog.showOpenDialog(null, {
-      properties: [ 'openDirectory' ],
-      title: 'フォルダ選択',
-      defaultPath: '.'
-    }, (dirNames) => {
-      console.log(dirNames);
-
-      this.ipcSend(dirNames[0]);
-    });
-  }
-
-  ipcSend(dirName) {
-    ipcRenderer.send('PICTURE-DIR:SEND', dirName);
-  }
-
-  ipcReply() {
-    ipcRenderer.on('PICTURE-DIR:REPLY:SUCCESS', (event, fullExifList) => {
-      console.log('ipcReply:', fullExifList);
-
-      fullExifList.forEach(({ fileName, exifData, filePath }, idx) => {
-        g.picturesList.setTableRow({
-          filePath,
-          fileName,
-          exifData,
-          fileNum: idx,
-        });
-      });
-    });
-
-    ipcRenderer.on('PICTURE-DIR:REPLY:ERROR', (event, err) => {
-      console.log('ipcReply error.');
-    });
-  }
-}
+const { clipboard } = require('electron');
 
 
 class PicturesList {
@@ -136,3 +81,5 @@ class PicturesList {
   }
 
 }
+
+module.exports = PicturesList;
